requirements.txt
python-telegram-bot==21.8
fastapi>=0.109.2
uvicorn>=0.27.1
sqlalchemy>=2.0.27
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
yt-dlp>=2024.3.10
python-multipart>=0.0.9
aiofiles>=23.2.1
pydantic>=2.6.1
Pillow>=10.2.0 